- removed duplicate interfaces - added interface for processed policy - fixed tests - added FRPolicy method to login failure
