+++ src/fr-policy/interfaces.ts
@@ -1,33 +1,10 @@
-interface ErrorResponse {
-  code: number;
-  reason: string;
-  message: string;
-  detail: ErrorDetail;
-}
+import { FailedPolicyRequirement } from '../auth/interfaces';
 
-interface ErrorDetail {
-  result: boolean;
-  failedPolicyRequirements: FailedPolicyRequirement[];
-}
-
-interface FailedPolicyRequirement {
-  policyRequirements: PolicyRequirement[];
-  property: string;
-}
-
-interface PolicyRequirement {
-  params?: any;
-  policyRequirement: string;
-}
-
-interface PolicyRequirementParams {
-  [key: string]: number;
+interface ProcessedPropertyError {
+  messages: string[];
+  detail: FailedPolicyRequirement;
 }
 
 export {
-  ErrorResponse,
-  ErrorDetail,
-  FailedPolicyRequirement,
-  PolicyRequirement,
-  PolicyRequirementParams,
+  ProcessedPropertyError,
 };
